requirements.txt
graphviz==0.20.3
matplotlib==3.9.2
neat_python==0.92
numpy==2.1.3
pygame==2.6.1
pytest==8.3.3
